bug(#275): move layout to componenets
